Add apb init to apb_directory class
